rename SqeTrainScheduler
