+++ python/fedml/simulation/mpi/fedavg_seq/FedAVGAggregator.py
@@ -9,8 +9,7 @@ import wandb
 
 from .utils import transform_list_to_tensor
 
-from ....core.schedule.scheduler import scheduler
-from ....core.schedule.scheduler import scheduler_c
+from ....core.schedule.seq_train_scheduler import SeqTrainScheduler
 from ....core.schedule.runtime_estimate import t_sample_fit
 
 
@@ -163,9 +162,9 @@ class FedAVGAggregator(object):
             workloads = np.array([ self.train_data_local_num_dict[client_id] for client_id in client_indexes])
             constraints = np.array([1]*self.worker_num)
             memory = np.array([100])
-            my_scheduler = scheduler_c(workloads, constraints, memory,
+            my_scheduler = SeqTrainScheduler(workloads, constraints, memory,
                 fit_funcs, uniform_client=True, uniform_gpu=False)
-            # my_scheduler = scheduler_c(workloads, constraints, memory, self.train_data_local_num_dict,
+            # my_scheduler = SeqTrainScheduler(workloads, constraints, memory, self.train_data_local_num_dict,
             #     fit_funcs, uniform_client=True, uniform_gpu=False)
             y_schedule, output_schedules = my_scheduler.DP_schedule(mode)
             client_schedule = []
